+++ src/app.js
@@ -2,6 +2,7 @@ import express from "express";
 import bodyParser from "body-parser";
 import { logger } from "./lib";
 import dictionaryRoutes from "./dictionary-routes";
+import productRoutes from "./product-routes";
 
 const app = express();
 
@@ -9,5 +10,6 @@ app.use(bodyParser.json());
 app.use(logger);
 app.use(express.static("./client"));
 app.use("/dictionary", dictionaryRoutes);
+app.use("/products", productRoutes);
 
 export default app;
